database/requirements.txt
numpy
pandas
requests
ipykernel
mysql-connector-python
